+++ Assets/Scripts/Enemies/EY/EY_AI.cs
@@ -16,13 +16,16 @@ public class EY_AI : EnemyAI
     public float TooCloseBeginChase;
     public float TooCloseEndChase;
 
+    public AudioClip LockOnSound;
+    public AudioClip ShootSound;
+
     private bool chase = false;
     private bool tooCloseChase = false;
     private bool canShoot = true;
 
     private void Update()
     {
-        if (BeginChase + 5f > DistanceBetweenPlayer() && canShoot)
+        if (BeginChase + 5.0f > DistanceBetweenPlayer() && canShoot)
         {
             StartCoroutine(BeginShot());
         }
@@ -49,24 +52,21 @@ public class EY_AI : EnemyAI
         {
             tooCloseChase = false;
         }
+        else if (TooCloseBeginChase > distance)
+        {
+            tooCloseChase = true;
+            aiPath.IsStopped = false;
+        }
+        else if (BeginChase < distance)
+        {
+            chase = true;
+            aiPath.IsStopped = false;
+        }
         else
         {
-            if (TooCloseBeginChase > distance)
-            {
-                tooCloseChase = true;
-                aiPath.IsStopped = false;
-            }
-            else if (BeginChase < distance)
-            {
-                chase = true;
-                aiPath.IsStopped = false;
-            }
-            else
-            {
-                chase = false;
-                tooCloseChase = false;
-                aiPath.IsStopped = true;
-            }
+            chase = false;
+            tooCloseChase = false;
+            aiPath.IsStopped = true;
         }
     }
 
@@ -75,9 +75,13 @@ public class EY_AI : EnemyAI
         // Attack separated in phases
         canShoot = false;
         laser.DrawLine = false;
-        yield return new WaitForSeconds(3f);
+        yield return new WaitForSeconds(3.0f);
+
+        audioSource.PlayOneShot(LockOnSound);
         laser.DrawLine = true;
-        yield return new WaitForSeconds(2f);
+        yield return new WaitForSeconds(2.0f);
+
+        audioSource.PlayOneShot(ShootSound);
         laser.DrawLine = false;
         canShoot = true;
         Fire();