+++ Assets/Scripts/Map/NodeBlueprint.cs
@@ -1,0 +1,11 @@
+﻿// Copyright (c) TigardHighGDC
+// SPDX-License SPDX-License-Identifier: Apache-2.0
+
+using UnityEngine;
+
+[CreateAssetMenu]
+public class NodeBlueprint : ScriptableObject
+{
+    public Sprite Sprite;
+    public NodeType NodeType;
+}